+++ src/ImathTest/testJacobiEigenSolver.cpp
@@ -62,10 +62,10 @@ template <typename TM>
 void
 verifyOrthonormal (const TM& A, const typename TM::BaseType threshold)
 {
-    const TM prod = A * A.transposed();
-    for (size_t i = 0; i < TM::dimensions(); ++i)
-        for (size_t j = 0; j < TM::dimensions(); ++j)
-            if (i == j) 
+    const TM prod = A * A.transposed ();
+    for (unsigned int i = 0; i < TM::dimensions (); ++i)
+        for (unsigned int j = 0; j < TM::dimensions (); ++j)
+            if (i == j)
                 assert (std::abs (prod[i][j] - 1) < threshold);
             else
                 assert (std::abs (prod[i][j]) < threshold);
@@ -78,9 +78,9 @@ computeThreshold(const TM& A)
    typedef typename TM::BaseType T;
    T maxAbsEntry(0);
 
-   for (size_t i = 0; i < TM::dimensions(); ++i)
-       for (size_t j = 0; j < TM::dimensions(); ++j)
-           maxAbsEntry = std::max (maxAbsEntry, std::abs(A[i][j]));
+    for (unsigned int i = 0; i < TM::dimensions (); ++i)
+        for (unsigned int j = 0; j < TM::dimensions (); ++j)
+            maxAbsEntry = std::max (maxAbsEntry, std::abs (A[i][j]));
 
    const T eps = std::numeric_limits<T>::epsilon();
    maxAbsEntry = std::max(maxAbsEntry, eps);
@@ -113,9 +113,9 @@ testJacobiEigenSolver(const TM& A)
 
     // Determinant of A and S
     TM MS;
-    for (size_t i = 0; i < TM::dimensions(); ++i)
-        for (size_t j = 0; j < TM::dimensions(); ++j)
-            if(i == j)
+    for (unsigned int i = 0; i < TM::dimensions (); ++i)
+        for (unsigned int j = 0; j < TM::dimensions (); ++j)
+            if (i == j)
                 MS[i][j] = S[i];
             else
                 MS[i][j] = 0;
@@ -125,9 +125,9 @@ testJacobiEigenSolver(const TM& A)
     // A = V * S * V^T
     TM MA = V * MS * V.transposed();
 
-    for (size_t i = 0; i < TM::dimensions(); ++i) 
-        for (size_t j =0; j < TM::dimensions(); ++j) 
-            assert(abs(A[i][j]-MA[i][j]) < threshold);
+    for (unsigned int i = 0; i < TM::dimensions (); ++i)
+        for (unsigned int j = 0; j < TM::dimensions (); ++j)
+            assert (abs (A[i][j] - MA[i][j]) < threshold);
 }
 
 template<class TM>
